pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "halos"
version = "0.1.0"
description = "HAL agent operating system - CLI tools for memory, jobs, cron, and backlog"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "feedparser>=6.0",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.27.0",
    "ipython>=9.10.0",
    "jinja2>=3.0",
    "onepassword-sdk>=0.4.0",
    "playwright>=1.58.0",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "youtube-transcript-api>=1.2.4",
]



[project.optional-dependencies]
graph = ["networkx>=3.0", "pyvis>=0.3.2"]
briefings = ["anthropic>=0.40.0"]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]
eventsource = ["nats-py>=2.9.0", "python-ulid>=3.0.0"]

[project.scripts]
memctl = "halos.memctl.cli:main"
nightctl = "halos.nightctl.cli:main"
cronctl = "halos.cronctl.cli:main"
logctl = "halos.logctl.cli:main"
reportctl = "halos.reportctl.cli:main"
agentctl = "halos.agentctl.cli:main"
hal-briefing = "halos.briefings.cli:main"
halctl = "halos.halctl.cli:main"
observe = "halos.eventsource.observe_cli:main"
halo-devlog = "halos.devlog_cli:main"
halo-telephony = "halos.telephony_cli:main"
trackctl = "halos.trackctl.cli:main"
dashctl = "halos.dashctl.cli:main"
mailctl = "halos.mailctl.cli:main"
calctl = "halos.calctl.cli:main"
statusctl = "halos.statusctl.cli:main"
backupctl = "halos.backupctl.cli:main"
ledgerctl = "halos.ledgerctl.cli:main"
docctl = "halos.docctl.cli:main"
watchctl = "halos.watchctl.cli:main"
secretctl = "halos.secretctl.cli:main"
drillctl = "halos.drillctl.cli:main"
journalctl = "halos.journalctl.cli:main"
advisorctl = "halos.advisorctl.cli:main"
hal = "halos.hal:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    'slow: marks tests as slow (deselect with -m "not slow")',
    "smoke: marks lightweight smoke tests",
    "fleet: requires live k8s cluster",
    "tier1: plumbing smoke (fast, post-deploy)",
    "tier2: NATS event flow verification",
    "tier3: chaos engineering (destructive)",
    "chaos: alias for tier3",
    "tier4: mock pipeline (cost-free integration)",
    "tier5: live Telegram liveness",
    "telegram: alias for tier5",
]

[tool.uv.workspace]
members = [
    "data/finance/ark-accounting",
    "data/advisors/guido/curriculum/scripts",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
